+++ api/RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## in develop
 
 * Change to make the user-agent string for dxScala more distinctive.
+* Support for collecting data objects / files nested inside JSON objects.
 
 ## 0.13.9 (2024-02-29)
 * adds `headJobOnDemand` attribute to jobNew call